feature: logic moved to userStore

+++ Frontend/src/stores/userStore.ts
@@ -1,38 +1,67 @@
 import { defineStore } from 'pinia'
-import type { User } from '../interfaces/types'
+import type {
+    PaginationState,
+    State,
+    User,
+    UserState
+} from '../interfaces/types'
 import axios from 'axios'
 
 export const useUserStore = defineStore('user', {
-    state: () => ({
-        user: null as User | null,
-        users: [] as User[],
-        isLoading: false,
-        showToast: false,
-        error: ''
+    state: (): {
+        user: UserState
+        pagination: PaginationState
+        state: State
+    } => ({
+        user: {
+            selected: null,
+            list: []
+        },
+        pagination: {
+            total: 0,
+            totalPages: 0,
+            currentPage: 1
+        },
+        state: {
+            isLoading: false,
+            showToast: false,
+            error: ''
+        }
     }),
     actions: {
         setUser(user: User) {
-            this.user = user
+            this.user.selected = user
+        },
+
+        async fetchUsers(page: number) {
+            this.state.isLoading = true
+            try {
+                const response = await axios.get(
+                    `http://localhost:4000/usuarios?page=${page}`
+                )
+                this.user.list = response.data.users
+                this.pagination.total = response.data.total
+                this.pagination.totalPages = response.data.totalPages
+                this.pagination.currentPage = response.data.page
+            } catch (err) {
+                this.state.error = 'Error al obtener los usuarios'
+            } finally {
+                this.state.isLoading = false
+            }
         },
 
         async deleteUser() {
-            this.isLoading = true
+            if (!this.user) return
             try {
                 await axios.delete(
-                    `http://localhost:4000/usuarios/${this.user?.id}`
+                    `http://localhost:4000/usuarios/${this.user.selected?.id}`
                 )
-                this.users = this.users.filter((u) => u.id !== this.user?.id)
-                return true
             } catch (error) {
-                this.error = 'Failed to delete user'
-                throw error
+                this.state.error = 'Failed to delete user'
             } finally {
-                this.user = null
-                this.showToast = true
-                this.isLoading = false
-                setTimeout(() => {
-                    this.showToast = false
-                }, 3000)
+                this.user.selected = null
+                this.state.showToast = true
+                this.fetchUsers(this.pagination.currentPage)
             }
         }
     }
